+++ src/main/java/com/br/mypoint/mypoint_skate/domain/repositories/RoleRepository.java
@@ -1,11 +1,12 @@
 package com.br.mypoint.mypoint_skate.domain.repositories;
 
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import com.br.mypoint.mypoint_skate.domain.models.Role;
 
 @Repository
-public interface RoleRepository extends JpaRepository<Role, Long> {
+public interface RoleRepository extends JpaRepository<Role, Integer> {
     
 }
